[refract] detele patch2md
