+++ src/targets/factory.py
@@ -1,9 +1,8 @@
 import os
+import git
 from abc import ABC, abstractmethod
 
-import git
-
-from patch2md import get_function_codes
+from tools import get_function_codes
 
 
 class TargetFactory(ABC):
